pachca-api-open-api-3-0-client/pyproject.toml
[tool.poetry]
name = "pachca-api-open-api-3-0-client"
version = "3.0.3"
description = "A client library for accessing PachcaAPI - OpenAPI 3.0"
authors = []
readme = "README.md"
packages = [
    {include = "pachca_api_open_api_3_0_client"},
]
include = ["CHANGELOG.md", "pachca_api_open_api_3_0_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.20.0,<0.28.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
